requirements.txt
Django==4.1.7
django-admin-sortable2==2.1.4
python-dotenv==1.0.0
Pillow==9.5.0